warning

+++ common/xqc_test.c
@@ -168,7 +168,7 @@ int test_pq(int argc, char* argv[])
 
     while (!xqc_pq_empty(&pq)) {
         xqc_pq_element_t* e = xqc_pq_top(&pq);
-        printf("element key:%lu\n", e->key);
+        printf("element key:%lu\n", (unsigned long)e->key);
         xqc_pq_pop(&pq);
     }
 
@@ -205,7 +205,7 @@ int test_pq(int argc, char* argv[])
 
     while (!xqc_pq_empty(&pq2)) {
         xqc_pq_item_t* e = (xqc_pq_item_t*)xqc_pq_top(&pq2);
-        printf("element key:%lu value:%d\n", e->key, *(int*)e->ptr);
+        printf("element key:%lu value:%d\n", (unsigned long)e->key, *(int*)e->ptr);
         xqc_pq_pop(&pq2);
     }
